post-receive.hook: Handle failure to find patch number
When pwclient info -h fails to come up with the number for the change in
question it will exit with a non-zero exit code.  This failure will
propagate upwards and exit the script there.  Make our call to
get_patch_id or in true so that our script here will see an empty id and
we continue on with the list.

Signed-off-by: Tom Rini <trini@konsulko.com>
Reviewed-by: Stephen Finucane <stephen@that.guru>