parser: Use a series even if the version differs
Currently, once we've found a series for a new submission, we check to
ensure that the version of that series matches that of the submission,
and we discard the series if not. The original intention for this was
given in commit 'c21b30525', where this was added:

  There are some things you probably shouldn't do on public mailing
  lists, but which people do anyway.

  - [PATCH 1/2] test: Add some lorem ipsum
    - [PATCH 2/2] test: Convert to Markdown
      - [PATCH v2 1/2] test: Add some lorem ipsum
        - [PATCH v2 2/2] test: Convert to Markdown

  We should correctly parse these...

This is unnecessary for two reasons:

- If the series is using proper references, then the mechanism that
  we use to prevent this issue with unversioned follow ups (also added
  in that patch) will work here: namely, we don't add submissions if an
  submission with a given number already exists in that series.
- If the series is not using references, we already have a check for
  version.

Seeing as this check is actually causing issues for legitimate typos, we
should just remove this check. Do so, adding a check to ensure we don't
regress.

Signed-off-by: Stephen Finucane <stephen@that.guru>
Closes-bug: #105