views: Fix "Add to bundle" dropdown on patch list view
The "Add to bundle" option in the patch list view requires the user's
list of bundles to be added to the context. Before PatchworkRequestContext
was removed, this was done in a context processor. When that was
refactored out, the bundles list was re-added in the patch detail view,
but not in the patch list view.

Add the bundles list in the patch list view to rectify this.

Reported-by: David Miller <davem@davemloft.net>
Reported-by: Michael Ellerman <mpe@ellerman.id.au>
Reported-by: Paul Mackerras <paulus@ozlabs.org>
Suggested-by: Michael Ellerman <mpe@ellerman.id.au>
Closes #116 ("Add to existing bundle drop down has disappeared from bottom of patch list page")
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Reviewed-by: Stephen Finucane <stephen@that.guru>
(cherry picked from commit b62c11755c21f2114a6f71aeb21a8c1e18617751)