Remove support for Django 1.8, 1.9, 1.10
These are now all EOL and Debian Testing supports Django 1.11 (LTS). We
can and should drop them.

This change does not remove the many compat wrappers. These will be
removed separately (there are a lot of them).

This leaves 1.11 as the only supported version. This will be remedied
shortly with the inclusion of Django 2.0 support.

Signed-off-by: Stephen Finucane <stephen@that.guru>
Signed-off-by: Daniel Axtens <dja@axtens.net>