Add support for Django 2.0
Nothing too complicated here except for the addition of a new compat
wrapper, which will be removed again shortly. According to the Django
release notes, Django should function with Python 3.4. However, it was
not possible to get this functioning due to the below error:

  Traceback (most recent call last):
    File ".../patchwork/manage.py", line 11, in <module>
    ...
    File ".../django/db/models/fields/related.py", line 313, in contribute_to_class
      'app_label': cls._meta.app_label.lower(),
  TypeError: unsupported operand type(s) for %: 'bytes' and 'dict'

This does not appear to be an issue with Patchwork but the exact root
cause has not been identified. As a result, only Python 3.5 and 3.6 are
marked as supported for this Django version.

As this is the first Python 3-only dependency we have, we need to start
making use of the 'python_version' environment marker.

Signed-off-by: Stephen Finucane <stephen@that.guru>
Signed-off-by: Daniel Axtens <dja@axtens.net>