travis: test against postgresql 10 and 11
postgresql 10 is the most recent major version, and it would be
good to keep an eye on postgres 11 compatibility also, so test
against both of those.

Specify PGPORT as a way to differentiate the new installations
from the postgres 9.6 server running on the default port (5432).
Use a local unix socket to work around the fact that by default
postgres only allows passwordless auth on local sockets - which
is tweaked for postgres 9.6 but not for later versions that aren't
in the usual Travis image.

Also print the current db version and user for validation.

Signed-off-by: Daniel Black <daniel@linux.ibm.com>
[cleanup, squash, commit message]
Signed-off-by: Daniel Axtens <dja@axtens.net>